Improved unit tests for samplers and fixed two minor bugs.
The unit tests for the ExchangeMC-derived classes in
test.cases.statistics.samplers have been greatly improved in terms of
actually failing if the algorithms don't work. Before, the target density
was just a normal distribution which does not require advanced
algorithms to sample from. It has now been replaced with a simple
multimodal distribution, which for low temperatures can only be
correctly sampled when employing some more advanced algorithm, e.g.
Replica Exchange. Sampling is assessed by comparing well occupation
numbers.
Also, for the HMCSampler class, there is now a test for a non-unity
mass matrix.
The tests may take a little longer now, but I prefer that over not noticing
bugs.

In the AlternatingAdjacentSwapScheme class of statistics.samplers.mc,
there was a minor bug: if there are only two replicae, a swap was attempted
only every second time swap_all() was called. This is fixed now.

In ThermostattedMDPropagator of statistics.samplers.mc.propagators,
an estimate of the number of random numbers to be drawn is calculated.
For very short trajectory lengths (like up to 5 steps), this number
could become negative. It is now set to an arbitrary minimum of 5.